Fix DM name display and stats grouping on dashboard
DM stats were grouped by event_dms.username which split the same DM's
stats when their Discord nickname changed between events. Now groups
by user_id with a LEFT JOIN to the players table for display names.

Extracted playerName formatter to dashboard/utils/playerName.js so it
can be used in both EJS templates and route handlers. Applied it to
all DM display points: home page top DMs, event detail, events list,
and DM stats (table + chart labels).
