+++ dashboard/views/dms.ejs
@@ -59,7 +59,7 @@
       <tbody>
         <% dmStats.forEach(dm => { %>
         <tr class="border-b border-gray-700/50 hover:bg-gray-700/30">
-          <td class="px-4 py-3 font-semibold"><%= dm.username %></td>
+          <td class="px-4 py-3 font-semibold"><%= playerName(dm.username, null) %></td>
           <td class="px-4 py-3 text-right text-sm text-gray-400"><%= dm.latest_start ? dm.latest_start.toISOString().split('T')[0] : '—' %></td>
           <td class="px-4 py-3 text-right text-sm text-gray-400"><%= dm.days_since_last_close != null ? dm.days_since_last_close : '—' %></td>
           <td class="px-4 py-3 text-right"><%= dm.events_started %></td>
