+++ dashboard/views/event-detail.ejs
@@ -49,7 +49,7 @@
         <ul class="space-y-2">
           <% event.dms.forEach(dm => { %>
           <li>
-            <span><%= dm.username || dm.user_id %></span>
+            <span><%= dm.username ? playerName(dm.username, null) : dm.user_id %></span>
           </li>
           <% }) %>
         </ul>
